OSv: ash fix process spawning
OSv or Linux build is selected with OSV_BUILD flag in include/busybox.h.

fork is ignored, and we pretend 0 is returned, so that child code does
get executed.
execve is replaced with osv_execve.
To wait on process termination, osv_waittid is used.

A few additional exit() calls have to be skipped.
A few additonal function do return in OSv, so remove NORETURN from their
declaration.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>