OSv ash: fix command not found error message when command was found
Running simple 'ls' in ash showed directory content. But at the end,
spurious error message 'command not found' was shown. This was due to
errno not being cleared by osv_execve. This was fixed in osv_execve, now
it clears errno on success, or sets it to meanigful value (currently
always ENOENT).

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>